backend/requirements.txt
amqp==5.3.1
anyio==4.6.2.post1
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
bioc==2.1
cbor==1.0.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
django-cors-headers==4.6.0
docopt==0.6.2
gensim==4.3.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.10
ijson==3.3.0
inscriptis==2.5.3
intervaltree==3.1.0
ir_datasets==0.5.9
jaro-winkler==2.0.3
joblib==1.4.2
jsonlines==4.0.0
kombu==5.4.2
Levenshtein==0.26.0
lxml==5.3.0
lz4==4.4.3
nest-asyncio==1.6.0
numpy==1.26.4
orcid==1.0.3
packaging==24.1
pandas==2.2.3
patsy==0.5.6
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyBioC @ git+https://github.com/OntoGene/PyBioC.git@6b29bd791fdcbc5c0cbef7a203ae86bb3600b47c
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
scikit-learn==1.5.2
scipy==1.13.1
semanticscholar==0.8.4
simplejson==3.19.3
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
statsmodels==0.14.4
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.66.5
trec-car-tools==2.6
tzdata==2024.2
unlzw3==0.2.3
urllib3==2.2.3
vine==5.1.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wcwidth==0.2.13
webencodings==0.5.1
wrapt==1.16.0
zlib-state==0.1.9

djangorestframework~=3.15.2